Fix duplicate entries being added to storage tag
Currently, if there is no player entity in the world, Dimord will add an
entry to the 'Book.tag.dimord' for each dimension tick, every tick. This
will persist until a player entity enters the world.

This commit fixes this behavior. The tag will be reset every tick until
a player entity enters the world (at which point the looping function
stops running).